added update size test
